Revert generic parameters
Attempted to add a generic parameter to ClassRowMapping class, but found
it to be time consuming and with little benefit. There were a lot of
type-safety warnings after the previous commit 226319f because I added a
generic parameter to MemberAccess which affected a couple other classes.
After a bit of reading of Java generics, and contemplating about adding
generics in the code, I decided to revert it back to the way it was.