Test multiple instances of same IN parameter
Execute a SELECT statement with an IN parameter that is used twice, one
in a WITH clause and another in the main SELECT query. But H2 does not
support non-recursive CTEs. Instead I put the subquery in an inline view
and that works.

- Remove README TODO regarding reflection API
- Create JdbcStatement method that returns a list of object arrays with
each array representing values of each row