Test ANSI printing checking SQL string equality
Tests in ANSISpec.groovy should follow this pattern: print the SQL
string of a SqlStatement object, then parse the SQL string using
SqlParser resulting in a identical SqlStatement object, then print the
new SqlStatement object, and finally assert that both SQL strings are
equal.

If the input SQL string is not the result of printing a SqlStatement
object, the SQL string will have to follow a certain convention in order
for the equality to be true, since the ANSI class is opinionated. For
example, in this commit I test parsing a SELECT statement based off of
the last example at
http://sqlity.net/en/1435/a-join-a-day-nested-joins/. In the example in
the link, "JOIN" is used instead of "INNER JOIN" (which ANSI class
prints out), and "AS" keyword is used for tables defined in the FROM
clause (ANSI class just prints the alias).

- README.md - Remove the following TODOs: "Adding JUnit tests using
Spock", "Conform to ANSI SQL", and "Implement nested JOINs"
- Refactor print(From) method in ANSI class to append comma character in
correct state
- Add Spock spec for testing ANSI class