Add method JdbcMapper.createQuery() by SQL string
The old benchmark for "testRetrievalSqlbuilderOrm" with a score of
4345.117 is now "testRetrievalSqlbOrmGetById" with a better score of
2671.766. The reason for the better score is probably because I have not
used the "install" Maven goal after commit 0115512. Changes in commit
226319f would not have been seen in the sql-builder-orm-benchmark
project. The commit improved the performance of JdbcMapper. A new
install of SqlbORM was required because the benchmark project was having
problems during Maven build.

I started using more of the new abbreviation for SqlbuilderORM, which is
SqlbORM.

The new query by ID tests for Hibernate and SqlbORM is about using HQL
or SQL respectively, to get an object by ID instead of using a ORM
method to get an entity by ID i.e. Query.get() or JdbcMapper.get().
These tests are a fairer comparision to Sql2o because Sql2o uses SQL
strings.

- Test query by ID for Hibernate and SqlbORM
- Add JdbcMappper method to create query by SQL string instead of
SqlStatement object