Add PredicateOperators: BETWEEN, EXISTS, IN, LIKE
ExistPredicate is created from DSL static factory method instead of
Predicand interface default method because no Predicand (such as
TableColumn, Literal, etc) should be on the left side of an EXISTS
predicate. The EXISTS predicate does not depend on another predicand.

The LIKE predicate does not need its own class vs BETWEEN, EXISTS, or IN
because the structure is similar to the comparison predicates (>, >=, =,
<, <=). The right side of BETWEEN requires two ValueExpressions. The
EXISTS predicate does not have a left side and only accepts a
SelectStatement object on the right. The IN predicate contains a
non-empty List of ValueExpressions.

- Implement BETWEEN with class BetweenPredicatre, and
Predicand.between() and Predicand.notBetween()
- Implement EXISTS with class ExistPredicate, and DSL.exists() and
DSL.notExists()
- Implement IN (...) with class InPredicate, and Predicand.in()
Predicand DSL.notIn()
- Implement LIKE with Predicand.like() and Predicand.notLike()
- Fix bug where passing a null to DSL.literal(Number) did not check if
the Number was null and call literalNull()
- Add missing set operation methods except(), intersect(), union(),
unionAll() in SelectStatement
- Add ValueExpression.ifNull() method to return LiteralNull.INSTANCE if
the ValueExpression parameter is null