Implement MERGE statement
Will not support <joined-table> for MERGE statement's source table. Only
<table name> and <table query> will be supported. Inner joins are
error-prone to selecting duplicate fields for the merge condition. With
<table query>, we can use DISTINCT to remove duplicate parent table
records, or instead of using an inner join, use EXISTS to check for the
existence of child table records.

- Create MergeAction class for MergeStatement to hold a list of WHEN
MATCHED clauses
- Create DMLStatement marker interface to designate DeleteStatement,
InsertStatement, MergeStatement, SelectStatement, and UpdateStatement as
DML statements. Useful for having a single field for a INSERT, DELETE,
or UPDATE in a MergeAction object
- Add shorthand for InsertStatement.DEFAULT as DSL.sqlDefault() method
- Change SelectColumn.ALL to TableColumn.ALL so that ColumnAlias.of()
can be applied e.g. ord.of(DSL.star()) which would result in "ord.*"
- Remove non-null table requirement in InsertStatement and
UpdateStatement constructors for MERGE statements
- Make SelectStatement implement TableReference so that inline views are
possible
- Add SelectStatement.as(TableAlias) method to set an alias for the
inline view in FROM clause or table source in MERGE statement