Test reading dsl.Expressions
dsl.Expression's left, right and operator fields had to be removed.
Having a binary tree where the left/right nodes were of the type
Expression resulted in a sqlbuilder.Expression with nested expressions
e.g. ((a || b) || c) when it should just be a || b || c. Those fields
were replaced by a List of ExpressionNode which is an interface with a
method that returns an int to determine if the ExpressionNode is an
expression operand, operator, or an expression.

- Test building different sqlbuilder.Expressions with string
concatenation and with arithmetic operators
- Refactor dsl.Expression by replacing left/right and operator fields
with List<ExpressionNode>, and override ExpressionBuilder
- Refactor building sqlbuilder.Expression in DSLTransformerdefault
methods to append new operations to the list
- Refactor ExpressionBuilder methods to group any right expressions, for
consistency with DSL.group(Expression)
- Return "this" reference when Condition and Expression is already
grouped, and when JoinedTable is already nested